# Refactor Aggregator for improve performance and prepare refactor transform_aggregator_final