--- a/query/src/pipelines/transforms/group_by/aggregator.rs
+++ b/query/src/pipelines/transforms/group_by/aggregator.rs
@@ -83,7 +83,8 @@ impl<Method: HashMethod + PolymorphicKeysHelper<Method>> Aggregator<Method> {
     }
 
     #[inline(always)]
-    fn execute(params: &AggregatorParams, block: &DataBlock, places: &[StateAddr]) -> Result<()> {
+    #[allow(clippy::ptr_arg)] // &[StateAddr] slower than &StateAddrs ~20%
+    fn execute(params: &AggregatorParams, block: &DataBlock, places: &StateAddrs) -> Result<()> {
         let aggregate_functions = &params.aggregate_functions;
         let offsets_aggregate_states = &params.offsets_aggregate_states;
         let aggregate_arguments_columns = Self::aggregate_arguments(block, params)?;