--- a/query/src/pipelines/transforms/group_by/aggregator_params.rs
+++ b/query/src/pipelines/transforms/group_by/aggregator_params.rs
@@ -16,13 +16,18 @@ use std::sync::Arc;
 
 use common_datavalues::DataSchemaRef;
 use common_exception::Result;
-use common_functions::aggregates::AggregateFunctionRef;
+use common_functions::aggregates::{AggregateFunctionRef, get_layout_offsets};
 use common_planners::Expression;
+use std::alloc::Layout;
 
 pub struct AggregatorParams {
     pub aggregate_functions: Vec<AggregateFunctionRef>,
     pub aggregate_functions_column_name: Vec<String>,
     pub aggregate_functions_arguments_name: Vec<Vec<String>>,
+
+    // about function state memory layout
+    pub layout: Layout,
+    pub offsets_aggregate_states: Vec<usize>,
 }
 
 pub type AggregatorParamsRef = Arc<AggregatorParams>;
@@ -39,10 +44,14 @@ impl AggregatorParams {
             aggregate_functions_arguments_name.push(expr.to_aggregate_function_names()?);
         }
 
+        let (states_layout, states_offsets) = unsafe { get_layout_offsets(&aggregate_functions) };
+
         Ok(Arc::new(AggregatorParams {
             aggregate_functions,
             aggregate_functions_column_name,
             aggregate_functions_arguments_name,
+            layout: states_layout,
+            offsets_aggregate_states: states_offsets,
         }))
     }
 }