I hereby agree to the terms of the CLA available at: https://datafuse.rs/policies/cla/

## Summary

- [x] Refactoring the aggregator can better adapt to refactor `transform_aggregator_final`
- [x] Investigate and fix the causes of performance degradation
  - [x] Inline not work
  - [x] Declare local references
  - [x] While match optimize match while
  - [x] clippy::ptr_arg

```
Query:
  SELECT  sum(number) FROM numbers_mt(10000000000)  group by number % 3, number % 4, number % 5;

Before:
  60 rows in set. Elapsed: 12.008 sec. Processed 10.00 billion rows, 80.00 GB (832.77 million rows/s., 6.66 GB/s.)

After:
  60 rows in set. Elapsed: 11.258 sec. Processed 10.00 billion rows, 80.00 GB (888.24 million rows/s., 7.11 GB/s.)

ClickHouse (master):
  60 rows in set. Elapsed: 12.152 sec. Processed 10.00 billion rows, 80.00 GB (822.93 million rows/s., 6.58 GB/s.)
```

## Changelog

- Performance Improvement